ajusta malloc em get_arg de hex_upper_format hex_lower_format pointer_format

+++ sources/formats/get_hex_upper_format.c
@@ -14,7 +14,7 @@
 
 static char	*formatter(void *data)
 {
-	unsigned long			*unsigned_data;
+	unsigned int			*unsigned_data;
 	t_hex_base_converter	*converter;
 	char					*result;
